+++ util/tracing/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Use `capture_fail!` to handle errors from the reporter thread.
 
 ## 0.2.0 - 2019-03-29
 ### Changed