+++ util/tracing/CHANGELOG.md
@@ -1,3 +1,4 @@
+<!-- markdownlint-disable MD022 MD024 MD032 -->
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -5,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.4.1 - 2022-09-12
+### Changed
+- Fix new warnings from clippy.
 
 ## 0.4.0 - 2020-05-28
 ### Changed